feat(init): configure line length during init Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
