+++ src/init/generation.rs
@@ -559,394 +559,8 @@ pub(super) fn prompt_config_dir(existing: Option<&str>, yes: bool) -> Result<Str
     }
 }
 
-/// Writes a skeleton `flint.toml` in `config_dir`. Creates the directory if needed.
-/// Returns `true` if the file was written, `false` if it already existed.
-///
-/// `exclude_managers`: when `Some`, populates `exclude_managers` in `[checks.renovate-deps]`
-/// with the given list (migrated from `RENOVATE_TRACKED_DEPS_EXCLUDE`). When `None` and
-/// `has_renovate` is true, writes a commented-out placeholder instead.
-pub(super) fn generate_flint_toml(
-    config_dir: &Path,
-    base_branch: &str,
-    has_renovate: bool,
-    exclude_managers: Option<&[String]>,
-) -> Result<bool> {
-    let toml_path = config_dir.join("flint.toml");
-    if toml_path.exists() {
-        return Ok(false);
-    }
-    std::fs::create_dir_all(config_dir)?;
-    let mut content = String::from("[settings]\n");
-    if base_branch != "main" {
-        content.push_str(&format!("base_branch = \"{base_branch}\"\n"));
-    }
-    content.push_str("# exclude = [\"CHANGELOG\\\\.md\"]\n");
-    if has_renovate {
-        content.push_str("\n[checks.renovate-deps]\n");
-        match exclude_managers {
-            Some(managers) if !managers.is_empty() => {
-                let list = managers
-                    .iter()
-                    .map(|m| format!("\"{m}\""))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                content.push_str(&format!("exclude_managers = [{list}]\n"));
-            }
-            _ => content.push_str("# exclude_managers = []\n"),
-        }
-    }
-    std::fs::write(&toml_path, &content)?;
-    println!("  wrote {}", toml_path.display());
-    Ok(true)
-}
-
-/// Generates `.rumdl.toml` in the project root when rumdl is being set up.
-/// Returns `true` if the file was written (or an older markdownlint variant was replaced).
-pub(super) fn generate_rumdl_config(project_root: &Path) -> Result<bool> {
-    const LEGACY_CONFIG_NAMES: &[&str] = &[
-        ".markdownlint.json",
-        ".markdownlint.jsonc",
-        ".markdownlint.yaml",
-        ".markdownlint.yml",
-        ".markdownlint-cli2.jsonc",
-        ".markdownlint-cli2.yaml",
-        ".markdownlint-cli2.yml",
-        ".markdownlint-cli2.cjs",
-        ".markdownlint-cli2.mjs",
-    ];
-    let target = project_root.join(".rumdl.toml");
-    if target.exists() {
-        return Ok(false);
-    }
-    for name in LEGACY_CONFIG_NAMES {
-        let legacy = project_root.join(name);
-        if legacy.exists() {
-            std::fs::remove_file(&legacy)?;
-            println!("  removed {} (replaced by .rumdl.toml)", legacy.display());
-        }
-    }
-    let content = concat!(
-        "[MD013]\n",
-        "enabled = true\n",
-        "line-length = 120\n",
-        "code-blocks = false\n",
-        "tables = false\n",
-    );
-    std::fs::write(&target, content)?;
-    println!("  wrote {}", target.display());
-    Ok(true)
-}
-
-/// Updates an existing editorconfig-checker config so Markdown is excluded when
-/// rumdl owns Markdown formatting and line-length enforcement.
-///
-/// Checks both the project root and `config_dir`, updating the first config
-/// file that exists. Returns `true` when a config was changed.
-pub(super) fn exclude_markdown_from_editorconfig_checker(
-    project_root: &Path,
-    config_dir: &Path,
-) -> Result<bool> {
-    const MARKDOWN_EXCLUDE: &str = ".*\\.md$";
-    let candidates = [
-        config_dir.join(".editorconfig-checker.json"),
-        project_root.join(".editorconfig-checker.json"),
-    ];
-
-    for path in candidates {
-        if !path.exists() {
-            continue;
-        }
-
-        let content = std::fs::read_to_string(&path)?;
-        let mut value: serde_json::Value = serde_json::from_str(&content)
-            .with_context(|| format!("failed to parse {}", path.display()))?;
-        let Some(obj) = value.as_object_mut() else {
-            anyhow::bail!("{} is not a JSON object", path.display());
-        };
-
-        let key = if obj.contains_key("Exclude") {
-            "Exclude"
-        } else if obj.contains_key("exclude") {
-            "exclude"
-        } else {
-            "Exclude"
-        };
-
-        let entry = obj
-            .entry(key.to_string())
-            .or_insert_with(|| serde_json::Value::Array(vec![]));
-        let Some(items) = entry.as_array_mut() else {
-            anyhow::bail!("{} field in {} is not an array", key, path.display());
-        };
-
-        if items
-            .iter()
-            .any(|v| v.as_str().is_some_and(|s| s == MARKDOWN_EXCLUDE))
-        {
-            return Ok(false);
-        }
-
-        items.push(serde_json::Value::String(MARKDOWN_EXCLUDE.to_string()));
-        let updated = serde_json::to_string_pretty(&value)? + "\n";
-        std::fs::write(&path, updated)?;
-        return Ok(true);
-    }
-
-    Ok(false)
-}
-
-/// Generates `.yamllint.yml` in the project root when yaml-lint is being set up.
-pub(super) fn generate_yamllint_config(project_root: &Path) -> Result<bool> {
-    let target = project_root.join(".yamllint.yml");
-    if target.exists() {
-        return Ok(false);
-    }
-    let content = concat!(
-        "extends: relaxed\n",
-        "\n",
-        "rules:\n",
-        "  document-start: disable\n",
-        "  line-length:\n",
-        "    max: 120\n",
-        "  indentation:\n",
-        "    spaces: 2\n",
-    );
-    std::fs::write(&target, content)?;
-    println!("  wrote {}", target.display());
-    Ok(true)
-}
-
-/// Generates `biome.json` in the project root when biome is being set up and no
-/// existing biome config is present.
-///
-/// Flint writes explicit space indentation to avoid Biome's default tab
-/// formatting surprising consumers during rollout.
-pub(super) fn generate_biome_config(project_root: &Path) -> Result<bool> {
-    const EXISTING_CONFIG_NAMES: &[&str] = &["biome.json", "biome.jsonc"];
-    if EXISTING_CONFIG_NAMES
-        .iter()
-        .map(|name| project_root.join(name))
-        .any(|path| path.exists())
-    {
-        return Ok(false);
-    }
-
-    let target = project_root.join("biome.json");
-    let content = [
-        "{",
-        "  \"formatter\": {",
-        "    \"indentStyle\": \"space\",",
-        "    \"indentWidth\": 2",
-        "  }",
-        "}",
-        "",
-    ]
-    .join("\n");
-    std::fs::write(&target, content)?;
-    println!("  wrote {}", target.display());
-    Ok(true)
-}
-
 /// Generates `.github/workflows/lint.yml` if it does not already exist.
 /// Returns `true` if the file was written.
-pub(super) fn generate_lint_workflow(
-    project_root: &Path,
-    base_branch: &str,
-    has_rust: bool,
-) -> Result<bool> {
-    let workflows_dir = project_root.join(".github/workflows");
-    let workflow_path = workflows_dir.join("lint.yml");
-    if workflow_path.exists() {
-        return Ok(false);
-    }
-    std::fs::create_dir_all(&workflows_dir)?;
-    let push_comment = if has_rust {
-        " # warms the Rust cache so PR branches get a cache hit"
-    } else {
-        ""
-    };
-    let rust_steps = if has_rust {
-        "\n      - uses: Swatinem/rust-cache@c19371144df3bb44fab255c43d04cbc2ab54d1c4 # v2.9.1\n\n      - name: Install Rust lint components\n        run: rustup component add clippy rustfmt\n"
-    } else {
-        ""
-    };
-    let content = format!(
-        r#"name: Lint
-
-on:
-  push:
-    branches: [{base_branch}]{push_comment}
-  pull_request:
-    branches: [{base_branch}]
-
-permissions:
-  contents: read
-
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-        with:
-          persist-credentials: false
-          fetch-depth: 0
-
-      - name: Setup mise
-        uses: jdx/mise-action@1648a7812b9aeae629881980618f079932869151 # v4.0.1
-        with:
-          version: v2026.4.18
-          sha256: 6ae2d5f0f23a2f2149bc5d9bf264fe0922a1da843f1903e453516c462b23cc1f
-{rust_steps}
-      - name: Lint
-        env:
-          GITHUB_TOKEN: ${{{{ github.token }}}}
-          GITHUB_HEAD_SHA: ${{{{ github.event.pull_request.head.sha }}}}
-        run: mise run lint
-"#
-    );
-    std::fs::write(&workflow_path, content)?;
-    println!("  wrote {}", workflow_path.display());
-    Ok(true)
-}
-
-/// Adds a `[tasks.<name>]` entry only when it is not already present.
-/// Returns `true` if an entry was added.
-fn add_task_if_absent(
-    tasks: &mut toml_edit::Table,
-    name: &str,
-    description: &str,
-    run: &str,
-) -> bool {
-    if tasks.contains_key(name) {
-        return false;
-    }
-    write_task(tasks, name, description, run);
-    true
-}
-
-/// Unconditionally writes a `[tasks.<name>]` entry (adds or replaces).
-fn write_task(tasks: &mut toml_edit::Table, name: &str, description: &str, run: &str) {
-    let mut t = toml_edit::Table::new();
-    t.insert("description", toml_edit::value(description));
-    t.insert("run", toml_edit::value(run));
-    tasks.insert(name, toml_edit::Item::Table(t));
-}
-
-/// Returns `true` when the named task has a `depends` array where at least one
-/// entry is in `removed_tasks`. Used to detect tasks made stale by v1 removal.
-fn task_has_removed_dep(tasks: &toml_edit::Table, name: &str, removed: &[String]) -> bool {
-    let Some(item) = tasks.get(name) else {
-        return false;
-    };
-    let Some(task) = item.as_table() else {
-        return false;
-    };
-    let Some(depends) = task.get("depends").and_then(|v| v.as_array()) else {
-        return false;
-    };
-    depends.iter().any(|v| {
-        v.as_str()
-            .map(|s| removed.iter().any(|r| r == s))
-            .unwrap_or(false)
-    })
-}
-
-/// Adds `[env] FLINT_CONFIG_DIR` and the standard `lint*` tasks to `mise.toml`,
-/// skipping any that are already present.
-///
-/// When `removed_v1_tasks` is non-empty, standard tasks whose `depends` reference
-/// any of those removed tasks are replaced (they became stale after v1 removal).
-///
-/// Returns `true` if the file was changed.
-pub(super) fn apply_env_and_tasks(
-    mise_path: &Path,
-    config_dir_rel: &str,
-    has_slow: bool,
-    removed_v1_tasks: &[String],
-) -> Result<bool> {
-    let content = std::fs::read_to_string(mise_path).unwrap_or_default();
-    let mut doc: toml_edit::DocumentMut = content
-        .parse()
-        .unwrap_or_else(|_| toml_edit::DocumentMut::new());
-    let mut changed = false;
-
-    // [env] — add FLINT_CONFIG_DIR if absent
-    {
-        if !doc.contains_key("env") {
-            doc.insert("env", toml_edit::Item::Table(toml_edit::Table::new()));
-        }
-        let env = doc["env"].as_table_mut().context("[env] is not a table")?;
-        if !env.contains_key("FLINT_CONFIG_DIR") {
-            env.insert("FLINT_CONFIG_DIR", toml_edit::value(config_dir_rel));
-            changed = true;
-        }
-    }
-
-    // [tasks] — add lint / lint:fix / (lint:pre-commit)
-    {
-        if !doc.contains_key("tasks") {
-            let mut tasks_table = toml_edit::Table::new();
-            tasks_table.set_implicit(true);
-            doc.insert("tasks", toml_edit::Item::Table(tasks_table));
-        }
-        let tasks = doc["tasks"]
-            .as_table_mut()
-            .context("[tasks] is not a table")?;
-
-        // Replace the lint task when it was made stale by v1 removal (its depends
-        // referenced removed tasks and would now fail). Otherwise add if absent.
-        let lint_stale = task_has_removed_dep(tasks, "lint", removed_v1_tasks);
-        if lint_stale {
-            write_task(tasks, "lint", "Run all lints", "flint run");
-            changed = true;
-        } else {
-            changed |= add_task_if_absent(tasks, "lint", "Run all lints", "flint run");
-        }
-
-        changed |= add_task_if_absent(tasks, "lint:fix", "Auto-fix lint issues", "flint run --fix");
-        if has_slow {
-            changed |= add_task_if_absent(
-                tasks,
-                "lint:pre-commit",
-                "Fast auto-fix lint (skips slow checks) — for pre-commit/pre-push hooks",
-                "flint run --fix --fast-only",
-            );
-        }
-    }
-
-    if changed {
-        std::fs::write(mise_path, doc.to_string())?;
-    }
-    Ok(changed)
-}
-
-/// Offers to install the git pre-commit hook via `flint hook install`.
-/// Prompts the user unless `yes` is true. Silently skips if the hook is already installed.
-pub(super) fn maybe_install_hook(project_root: &Path, yes: bool) -> Result<()> {
-    let hook_path = project_root.join(".git/hooks/pre-commit");
-    if hook_path.exists() {
-        return Ok(());
-    }
-
-    let install = if yes {
-        true
-    } else {
-        print!(
-            "Install pre-commit hook (runs `flint run --fix --fast-only` before each commit)? [Y/n] "
-        );
-        io::stdout().flush()?;
-        let mut input = String::new();
-        io::stdin().lock().read_line(&mut input)?;
-        !input.trim().eq_ignore_ascii_case("n")
-    };
-
-    if install {
-        crate::hook::install(project_root)?;
-    }
-    Ok(())
-}
-
 #[cfg(test)]
 mod node_prereq_tests {
     use super::{ensure_node_for_npm, needs_node_for_npm};
