+++ tests/cases/general/init-idempotent/files/.editorconfig
@@ -5,5 +5,6 @@ charset = utf-8
 end_of_line = lf
 indent_style = space
 indent_size = 2
+max_line_length = 120
 insert_final_newline = true
 trim_trailing_whitespace = true
